Sets up Xcodegen and tests
Fixes #1 - Create an Xcode project file with Xcodegen

Fixes #4 - Set up tests

This commit creates an Xcodegen project.yml file that is (mostly) properly configured to generate an xcodeproj file with three targets: OTPKit, OTPKitDemo, and OTPKitTests. I created a mostly non-functional class in the OTPKit framework as a proof of concept that the tests, framework, and demo app are all working properly.