Present sheets from the body view instead of a child view
Fixes #26

I don't have a strong enough grasp of SwiftUI to understand why it doesn't like it when we present a modal sheet from a child View yet, but I can tell empirically that this resolves the issue of modals appearing and then disappearing immediately. The `Attempt to present while a presentation is in progress` errors also disappear after this change.